# packages/contracts/openapi/occ-api.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "serviceApiKey": { "in": "header", "name": "x-api-key", "type": "apiKey" },
      "sessionCookie": { "in": "cookie", "name": "openclaw_occ.session_token", "type": "apiKey" }
    }
  },
  "info": { "title": "Development OCC API", "version": "0.1.0" },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/service-keys": {
      "post": {
        "description": "Requires a session or Installation-scoped service key with administer on the Installation. Issues a Better Auth key for an existing non-Agent ServicePrincipal in its exact scope; creates no identity or IAM grant. The plaintext key is returned only here.",
        "operationId": "createServiceKey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "expiresIn": {
                    "description": "Lifetime in seconds; defaults to 30 days.",
                    "maximum": 31536000,
                    "minimum": 86400,
                    "type": "integer"
                  },
                  "name": { "maxLength": 32, "minLength": 1, "pattern": "\\S", "type": "string" },
                  "namespaceId": {
                    "pattern": "^ns_[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                    "type": "string"
                  },
                  "servicePrincipalId": { "maxLength": 200, "minLength": 1, "type": "string" }
                },
                "required": ["servicePrincipalId", "name"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Created",
                  "properties": {
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "expiresAt": { "format": "date-time", "type": "string" },
                        "id": { "type": "string" },
                        "key": { "type": "string" },
                        "name": { "type": "string" },
                        "namespaceId": { "type": "string" },
                        "servicePrincipalId": { "type": "string" }
                      },
                      "required": ["id", "servicePrincipalId", "name", "expiresAt", "key"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["data", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Bad Request",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Unauthorized",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Forbidden",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Not Found",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Conflict",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Service Unavailable",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [{ "sessionCookie": [] }, { "serviceApiKey": [] }],
        "summary": "Issue a service API key",
        "tags": ["Authentication"],
        "x-openclaw-permissions": [
          { "action": "administer", "resourceKind": "installation", "scope": "requested" }
        ]
      }
    },
...
    "/api/auth/sign-out": {
      "post": {
        "description": "Revokes the current Better Auth session cookie.",
        "operationId": "signOut",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "OK",
                  "properties": {
                    "data": { "additionalProperties": true, "type": "object" },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["data", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Unauthorized",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Unauthorized"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "Service Unavailable",
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": { "type": "string" },
                        "message": { "type": "string" }
                      },
                      "required": ["code", "message"],
                      "type": "object"
                    },
                    "meta": {
                      "additionalProperties": false,
                      "properties": { "requestId": { "type": "string" } },
                      "required": ["requestId"],
                      "type": "object"
                    }
                  },
                  "required": ["error", "meta"],
                  "type": "object"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [{ "sessionCookie": [] }],
        "summary": "Sign out of the current session",
        "tags": ["Authentication"]
      }
    },
...
  },
  "security": [{ "sessionCookie": [] }, { "serviceApiKey": [] }]
}